internal/controllers/middlewares/auth/authMiddleware.go
package authMiddleware

import (
	"errors"
	"fmt"
	"github.com/gin-gonic/gin"
	"github.com/gmaschi/go-recipes-book/pkg/auth/tokenAuth"
	"github.com/gmaschi/go-recipes-book/pkg/tools/parseErrors"
	"net/http"
	"strings"
)

const (
	AuthorizationHeaderKey  = "authorization"
	AuthorizationTypeBearer = "bearer"
	AuthorizationPayloadKey = "authorization_payload"
)

func AuthMiddleware(tokenMaker tokenAuth.Maker) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		authorizationHeader := ctx.GetHeader(AuthorizationHeaderKey)
		if len(authorizationHeader) == 0 {
			err := errors.New("authorization not provided")
			ctx.AbortWithStatusJSON(http.StatusUnauthorized, parseErrors.ErrorResponse(err))
			return
		}

		fields := strings.Fields(authorizationHeader)
		if len(fields) < 2 {
			err := errors.New("invalid authorization header format")
			ctx.AbortWithStatusJSON(http.StatusUnauthorized, parseErrors.ErrorResponse(err))
			return
		}

		authorizationType := strings.ToLower(fields[0])
		if authorizationType != AuthorizationTypeBearer {
			err := fmt.Errorf("unsupported authorization format %s", authorizationType)
			ctx.AbortWithStatusJSON(http.StatusUnauthorized, parseErrors.ErrorResponse(err))
			return
		}

		accessToken := fields[1]

		payload, err := tokenMaker.VerifyToken(accessToken)
		if err != nil {
			ctx.AbortWithStatusJSON(http.StatusUnauthorized, parseErrors.ErrorResponse(err))
			return
		}

		ctx.Set(AuthorizationPayloadKey, payload)
		ctx.Next()
	}
}
